requirements.txt
yfinance>=0.2.0
pandas>=1.5.0
fredapi>=0.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0
python-dotenv>=0.19.0
plotly>=5.0.0